OCR_INE/requirements.txt
fastapi>=0.104
uvicorn[standard]>=0.24
python-multipart>=0.0.6
opencv-python-headless>=4.8
pytesseract>=0.3.10
numpy>=1.24
pydantic>=2.5
pydantic-settings>=2.1
Pillow>=10.0
pytest>=7.4
anyio>=4.0
pytest-anyio>=0.0.0
httpx>=0.25
